# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LINKY API — Pay-to-Speak on Base
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PlatformConfig {
  chainId: number;
  chainName: string;
  usdcAddress: string;
  /**
     * Deployed LinkyEscrow contract address; null until the operator has set LINKY_ESCROW_ADDRESS
     * @nullable
     */
  escrowAddress: string | null;
  treasuryAddress: string;
  platformFeeBps: number;
  /** Seconds after booking before the buyer can refund */
  refundDelaySeconds: number;
  explorerBaseUrl: string;
}

export interface NonceInput {
  wallet: string;
}

export interface NonceResponse {
  nonce: string;
  expiresAt: number;
}

export interface LoginInput {
  wallet: string;
  signature: string;
  timestamp: number;
  nonce: string;
}

export interface LoginResponse {
  token: string;
  expiresAt: number;
  wallet: string;
}

export type Channel = typeof Channel[keyof typeof Channel];


export const Channel = {
  messaging: 'messaging',
  calls: 'calls',
} as const;

export type MessagingPlatform = typeof MessagingPlatform[keyof typeof MessagingPlatform];


export const MessagingPlatform = {
  whatsapp: 'whatsapp',
  telegram: 'telegram',
} as const;

/**
 * Public-safe view of a seller's messaging channel — no raw handle.
 */
export interface MessagingChannel {
  enabled: boolean;
  platform: MessagingPlatform;
  /** USDC amount as a decimal string */
  priceUsdc: string;
}

/**
 * Public-safe view of a seller's calls channel — no raw booking URL.
 */
export interface CallsChannel {
  enabled: boolean;
  /** USDC amount as a decimal string */
  priceUsdc: string;
}

export interface MessagingChannelInput {
  enabled: boolean;
  platform?: MessagingPlatform;
  /** WhatsApp phone number (E.164) or Telegram username */
  handle?: string;
  priceUsdc?: string;
}

export interface CallsChannelInput {
  enabled: boolean;
  bookingUrl?: string;
  priceUsdc?: string;
}

/**
 * Public expert profile — never includes raw contact details for paid channels.
 */
export interface Expert {
  id: string;
  handle: string;
  wallet: string;
  name: string;
  /** Profession / one-liner shown publicly */
  about: string;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  xUrl?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  messaging?: MessagingChannel | null;
  calls?: CallsChannel | null;
  createdAt: string;
}

export interface ExpertInput {
  /**
     * @minLength 2
     * @maxLength 32
     * @pattern ^[a-z0-9-]+$
     */
  handle: string;
  wallet: string;
  signature: string;
  timestamp: number;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  about: string;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  xUrl?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  messaging?: MessagingChannelInput | null;
  calls?: CallsChannelInput | null;
}

export interface ExpertStats {
  handle: string;
  totalEarnedUsdc: string;
  totalBookings: number;
  completedBookings: number;
  messagingBookings: number;
  callsBookings: number;
}

export interface RevealedContact {
  channel: Channel;
  platform?: MessagingPlatform | null;
  /**
     * WhatsApp number or Telegram username (only present for messaging)
     * @nullable
     */
  handle?: string | null;
  /**
     * Calendar booking URL (only present for calls)
     * @nullable
     */
  bookingUrl?: string | null;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  paid: 'paid',
  completed: 'completed',
  refunded: 'refunded',
} as const;

/**
 * Where the booking originated.
 */
export type BookingSource = typeof BookingSource[keyof typeof BookingSource];


export const BookingSource = {
  web: 'web',
  agent: 'agent',
} as const;

export interface Booking {
  id: string;
  expertHandle: string;
  expertWallet: string;
  buyerWallet: string;
  /** @nullable */
  buyerEmail?: string | null;
  amountUsdc: string;
  feeUsdc: string;
  escrowBookingId: string;
  bookTxHash: string;
  /** @nullable */
  releaseTxHash?: string | null;
  /** @nullable */
  refundTxHash?: string | null;
  status: BookingStatus;
  channel: Channel;
  /** 'agent' for bookings created via the Agent API, 'web' (or null) for human bookings. */
  source?: BookingSource | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  releasedAt?: string | null;
  /** @nullable */
  refundedAt?: string | null;
}

export interface BookingInput {
  expertHandle: string;
  channel: Channel;
  /** @pattern ^0x[0-9a-fA-F]{64}$ */
  escrowBookingId: string;
  /** @pattern ^0x[0-9a-fA-F]{64}$ */
  bookTxHash: string;
  /** Buyer email; the unlocked contact is delivered here. */
  email: string;
}

/**
 * A booking plus the revealed contact, returned right after payment.
 */
export type BookingWithReveal = Booking & ({
  reveal?: RevealedContact | null;
  /** Whether the contact email was dispatched successfully. */
  emailSent?: boolean;
});

export interface CompleteBookingInput {
  /** @pattern ^0x[0-9a-fA-F]{64}$ */
  releaseTxHash: string;
}

export interface RefundBookingInput {
  /** @pattern ^0x[0-9a-fA-F]{64}$ */
  refundTxHash: string;
}

export interface PlatformStats {
  totalVolumeUsdc: string;
  /** Knowledge Transmitted — count of paid+completed bookings only. */
  totalBookings: number;
  totalExperts: number;
  /** Links Generated — sum of messaging_enabled and calls_enabled channels across all experts. */
  totalLinks: number;
}

export type ListExpertsParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};