# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LINKY API — Pay-to-Speak on Base
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /experts:
    get:
      operationId: listExperts
      tags: [experts]
      summary: List recent experts
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expert"
    post:
      operationId: createExpert
      tags: [experts]
      summary: Create or claim an expert profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpertInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expert"
        "409":
          description: Handle or wallet already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /experts/{handle}:
    get:
      operationId: getExpert
      tags: [experts]
      summary: Public expert profile
      parameters:
        - in: path
          name: handle
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expert"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /experts/by-wallet/{wallet}:
    get:
      operationId: getExpertByWallet
      tags: [experts]
      summary: Lookup the expert profile owned by a wallet
      parameters:
        - in: path
          name: wallet
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expert"
        "404":
          description: No profile for this wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /experts/{handle}/stats:
    get:
      operationId: getExpertStats
      tags: [stats]
      summary: Headline metrics for an expert profile
      parameters:
        - in: path
          name: handle
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpertStats"
  /experts/{handle}/reveal/{channel}:
    get:
      operationId: revealExpertContact
      tags: [experts]
      summary: Reveal seller contact for a channel (buyer-only)
      description: |
        Returns the seller's contact for a given channel iff the calling
        wallet has a confirmed paid booking for this expert + channel.
        Otherwise 403.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: handle
          required: true
          schema: { type: string }
        - in: path
          name: channel
          required: true
          schema:
            $ref: "#/components/schemas/Channel"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevealedContact"
        "403":
          description: Caller has not paid for this channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: linky-session
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Channel:
      type: string
      enum: [messaging, calls]
    MessagingPlatform:
      type: string
      enum: [whatsapp, telegram]
    MessagingChannel:
      type: object
      description: Public-safe view of a seller's messaging channel — no raw handle.
      required: [enabled, platform, priceUsdc]
      properties:
        enabled: { type: boolean }
        platform:
          $ref: "#/components/schemas/MessagingPlatform"
        priceUsdc:
          type: string
          description: USDC amount as a decimal string
    CallsChannel:
      type: object
      description: Public-safe view of a seller's calls channel — no raw booking URL.
      required: [enabled, priceUsdc]
      properties:
        enabled: { type: boolean }
        priceUsdc:
          type: string
          description: USDC amount as a decimal string
    MessagingChannelInput:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
        platform:
          $ref: "#/components/schemas/MessagingPlatform"
        handle:
          type: string
          description: WhatsApp phone number (E.164) or Telegram username
        priceUsdc: { type: string }
    CallsChannelInput:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
        bookingUrl: { type: string }
        priceUsdc: { type: string }
    Expert:
      type: object
      description: Public expert profile — never includes raw contact details for paid channels.
      required:
        - id
        - handle
        - wallet
        - name
        - about
        - createdAt
      properties:
        id: { type: string }
        handle: { type: string }
        wallet: { type: string }
        name: { type: string }
        about:
          type: string
          description: Profession / one-liner shown publicly
        linkedinUrl:
          type: ["string", "null"]
        xUrl:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        messaging:
          oneOf:
            - $ref: "#/components/schemas/MessagingChannel"
            - type: "null"
        calls:
          oneOf:
            - $ref: "#/components/schemas/CallsChannel"
            - type: "null"
        createdAt: { type: string }
    ExpertInput:
      type: object
      required:
        - handle
        - wallet
        - signature
        - timestamp
        - name
        - about
      properties:
        handle:
          type: string
          minLength: 2
          maxLength: 32
          pattern: "^[a-z0-9-]+$"
        wallet: { type: string }
        signature: { type: string }
        timestamp: { type: integer }
        name: { type: string, minLength: 1 }
        about: { type: string, minLength: 1 }
        linkedinUrl:
          type: ["string", "null"]
        xUrl:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        messaging:
          oneOf:
            - $ref: "#/components/schemas/MessagingChannelInput"
            - type: "null"
        calls:
          oneOf:
            - $ref: "#/components/schemas/CallsChannelInput"
            - type: "null"
    ExpertStats:
      type: object
      required: [handle, totalEarnedUsdc, totalBookings, completedBookings, messagingBookings, callsBookings]
      properties:
        handle: { type: string }
        totalEarnedUsdc: { type: string }
        totalBookings: { type: integer }
        completedBookings: { type: integer }
        messagingBookings: { type: integer }
        callsBookings: { type: integer }
    RevealedContact:
      type: object
      required: [channel]
      properties:
        channel:
          $ref: "#/components/schemas/Channel"
        platform:
          oneOf:
            - $ref: "#/components/schemas/MessagingPlatform"
            - type: "null"
        handle:
          type: ["string", "null"]
          description: WhatsApp number or Telegram username (only present for messaging)
        bookingUrl:
          type: ["string", "null"]
          description: Calendar booking URL (only present for calls)
...