# lib/api-zod/src/generated/types/callsChannelInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LINKY API — Pay-to-Speak on Base
 * OpenAPI spec version: 0.1.0
 */

export interface CallsChannelInput {
  enabled: boolean;
  bookingUrl?: string;
  priceUsdc?: string;
}