# lib/api-zod/src/generated/types/messagingChannel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LINKY API — Pay-to-Speak on Base
 * OpenAPI spec version: 0.1.0
 */
import type { MessagingPlatform } from './messagingPlatform';

/**
 * Public-safe view of a seller's messaging channel — no raw handle.
 */
export interface MessagingChannel {
  enabled: boolean;
  platform: MessagingPlatform;
  /** USDC amount as a decimal string */
  priceUsdc: string;
}